tests/codec/test_protocol.py
"""
Test the 'codec.protocol' module.
"""

# module under test
from runtimepy.codec.protocol import Protocol
from runtimepy.enum.registry import EnumRegistry

# internal
from tests.resources import resource


def test_protocol_basic():
    """Test basic interactions with protocol objects."""

    proto = Protocol(
        EnumRegistry.decode(resource("enums", "sample_enum.yaml"))
    )

    proto.add_field("test1", "uint8")

    # Add bit fields.
    with proto.add_bit_fields() as fields:
        fields.flag("flag1", enum="bool1")
        fields.flag("flag2")

    proto.add_field("test2", "uint8", enum="int1")

    proto["test2"] = "b"
    assert proto["test2"] == "b"

    proto["flag1"] = "valve_open"
    assert proto["flag1"] == "valve_open"

    proto["flag2"] = True
    assert proto["flag2"] is True
